requirements.txt
absl-py==1.4.0
aio-pika==8.2.3
aiofiles==23.2.1
aiogram==2.15
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiormq==6.4.2
aiosignal==1.3.1
APScheduler==3.9.1.post1
astunparse==1.6.3
async-timeout==4.0.3
attrs==22.1.0
Babel==2.14.0
bidict==0.23.1
blis==0.7.11
boto3==1.34.91
botocore==1.34.91
CacheControl==0.12.14
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.2.1
confection==0.1.4
confluent-kafka==2.3.0
cryptography==42.0.5
cycler==0.12.1
cymem==2.0.8
dask==2022.10.2
dnspython==2.3.0
docopt==0.6.2
fbmessenger==6.0.0
filelock==3.13.4
fire==0.6.0
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
future==1.0.0
gast==0.4.0
google-auth==2.29.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.62.2
h11==0.14.0
h5py==3.11.0
httptools==0.6.1
huggingface-hub==0.22.2
humanfriendly==10.0
idna==3.7
jax==0.4.26
jieba==0.42.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.2.0
jsonpickle==3.0.4
jsonschema==4.17.3
keras==2.12.0
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
libclang==18.1.1
locket==1.0.0
marisa-trie==1.1.0
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.5.3
mattermostwrapper==2.2
ml-dtypes==0.4.0
msgpack==1.0.8
multidict==5.2.0
murmurhash==1.0.10
networkx==2.6.3
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==20.9
pamqp==3.2.1
partd==1.4.1
pathlib_abc==0.1.1
pathy==0.11.0
pillow==10.3.0
pluggy==1.5.0
portalocker==2.8.2
preshed==3.0.9
prompt-toolkit==3.0.28
protobuf==4.23.3
psycopg2-binary==2.9.9
pyaml==24.4.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==1.10.9
pydot==1.4.2
PyJWT==2.8.0
pykwalify==1.8.0
pymongo==4.3.3
pyparsing==3.1.2
pyrsistent==0.20.0
python-crfsuite==0.9.10
python-dateutil==2.8.2
python-engineio==4.9.0
python-socketio==5.11.2
pytz==2022.7.1
PyYAML==6.0.1
questionary==1.10.0
randomname==0.1.5
rasa==3.6.20
rasa-sdk==3.6.2
redis==4.6.0
regex==2022.10.31
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rocketchat-API==1.30.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.8
s3transfer==0.10.1
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-learn==1.1.3
scipy==1.10.1
sentencepiece==0.1.99
sentry-sdk==1.14.0
simple-websocket==1.0.0
six==1.16.0
sklearn-crfsuite==0.3.6
slack_sdk==3.27.1
smart-open==6.4.0
spacy==3.4.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==1.4.52
srsly==2.4.8
structlog==23.3.0
structlog-sentry==2.1.0
tabulate==0.9.0
tarsafe==0.0.4
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-hub==0.13.0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-text==2.12.0
termcolor==2.4.0
terminaltables==3.1.10
thinc==8.1.12
threadpoolctl==3.4.0
tokenizers==0.13.3
toolz==0.12.1
tqdm==4.66.2
transformers==4.26.0
twilio==8.2.2
typer==0.7.0
typing-utils==0.1.0
typing_extensions==4.11.0
tzlocal==5.2
ujson==5.9.0
urllib3==2.2.1
uvloop==0.19.0
wasabi==0.10.1
wcwidth==0.2.13
webexteamssdk==1.6.1
websockets==10.4
Werkzeug==3.0.2
wrapt==1.14.1
wsproto==1.2.0
yarl==1.9.4
